tehtävä 4.21*

+++ osa4/blogilista/controllers/blogs.js
@@ -52,8 +52,23 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: "token invalid" });
+    }
+
+    const userid = decodedToken.id;
+    console.log("userid", userid);
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() === userid.toString()) {
+        await Blog.findByIdAndDelete(request.params.id);
+        response.status(204).end();
+    } else {
+        response
+            .status(401)
+            .json({ error: "user not authorized to delete this blog" });
+    }
 });
 
 blogsRouter.put("/:id", async (request, response, next) => {
